Добавлено модульное тестирование клиента и сервера

+++ server/src/main/java/ru/hj77/server/dto/CardDTO.java
@@ -6,7 +6,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 @Getter
 public class CardDTO {
-    final int id_card;
+    final Long id_card;
     final int pinCode;
     final double balance;
 }
